requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
SQLAlchemy==2.0.40
python-multipart==0.0.20
jinja2==3.1.6
openpyxl==3.1.5
itsdangerous==2.2.0
python-docx==1.1.2
